+++ bot/classes/teams.py
@@ -10,11 +10,27 @@ class Team:
         self.__name = name
         self.__players = list()
         self.__score = 0
+        self.__net = 0
         self.__deaths = 0
         self.__kills = 0
         self.__faction = 0
         self.__cap = 0
         self.__match = match
+
+    def getData(self):
+        playersData = list()
+        for p in self.__players:
+            playersData.append(p.getData())
+        data = {"name": self.__name,
+                "faction_id": self.__faction,
+                "score": self.__score,
+                "net": self.__net,
+                "deaths": self.deaths,
+                "kills": self.__kills,
+                "cap_points": self.__cap,
+                "players": playersData
+                }
+        return data
 
     @property
     def id(self):
@@ -39,6 +55,10 @@ class Team:
     @property
     def score(self):
         return self.__score
+
+    @property
+    def net(self):
+        return self.__net
 
     @property
     def cap(self):
@@ -78,9 +98,13 @@ class Team:
     def addCap(self, points):
         self.__cap += points
         self.__score += points
+        self.__net += points
 
     def addScore(self, points):
         self.__score += points
+
+    def addNet(self, points):
+        self.__net += points
 
     def addOneKill(self):
         self.__kills +=1
@@ -99,3 +123,10 @@ class Team:
     def onMatchReady(self):
         for p in self.__players:
             p.onMatchReady()
+
+    def onPlayerSub(self, subbed, newPlayer):
+        i = 0
+        while self.__players[i] is not subbed:
+            i+=1
+        active = type(subbed)(newPlayer, self)
+        self.__players[i] = active