+++ bot/modules/config.py
@@ -85,7 +85,8 @@ facilitiy_suffix = {
 _collections = {
     "users": "",
     "sBases": "",
-    "sWeapons" : ""
+    "sWeapons" : "",
+    "matches" : ""
 }
 
 database = {
@@ -120,11 +121,11 @@ def getConfig(file):
             _errorIncorrect(key, 'General', file)
 
     # Testing api key
-    url = f"http://census.daybreakgames.com/s:{general['api_key']}/get/ps2:v2/faction"
-    jdata = loads(get(url).content)
-    if 'error' in jdata:
-        raise ConfigError(
-            f"Incorrect api key: {general['api_key']} in '{file}'")
+    # url = f"http://census.daybreakgames.com/s:{general['api_key']}/get/ps2:v2/faction"
+    # jdata = loads(get(url).content)
+    # if 'error' in jdata:
+    #     raise ConfigError(
+    #         f"Incorrect api key: {general['api_key']} in '{file}'")
 
     # Channels section
     _checkSection(config, "Channels", file)