+++ bot/modules/display.py
@@ -70,11 +70,11 @@ async def remReaction(message, user=None):
     if user is None:
         global _client
         user = _client.user
-    await message.remove_reaction("✅", user)
+    return await message.remove_reaction("✅", user)
 
 async def imageSend(channelId, imagePath):
     channel = _client.get_channel(channelId)
-    await channel.send(file=File(imagePath))
+    return await channel.send(file=File(imagePath))
 
 ## PRIVATE:
 
@@ -88,23 +88,24 @@ def _registerHelp(msg):
         title='How to register?',
         description=f'You have to accept the rules in <#{cfg.channels["rules"]}> to register'
     )
-    embed.add_field(name='If you don\'t have a Jaeger account',
-                    value='`=r no account`\n',
-                    inline=False)
-    embed.add_field(name='If you have a Jaeger account',
-                    value='`=r charName` - If your character names have faction suffixes\n'
-                          '`=r charName1 charName2 charName3` - If your character names don\'t have faction suffixes'
-                    , inline=False)
-    embed.add_field(name='Notify feature',
-                    value='`=notify` - To join or leave the Notify feature\n'
-                          f'When suscribed to Notify, you can be mentionned with <@&{cfg.roles["notify"]}> when the queue is almost full'
-                    , inline=False)
+    embed.add_field(name  = 'If you don\'t have a Jaeger account',
+                    value = '`=r no account`\n',
+                    inline = False)
+    embed.add_field(name  = 'If you have a Jaeger account',
+                    value = '`=r charName` - If your character names have faction suffixes\n'
+                            '`=r charName1 charName2 charName3` - If your character names don\'t have faction suffixes',
+                    inline = False)
+    embed.add_field(name  = 'Notify feature',
+                    value = '`=notify` - To join or leave the Notify feature\n'
+                            f'When suscribed to Notify, you can be mentionned with <@&{cfg.roles["notify"]}> '
+                            'when the queue is almost full',
+                    inline = False)
     try:
         if isAdmin(msg.author):
-            embed.add_field(name="Staff Commands",
-                            value='`=unregister @player` - Permanently remove player profile from the system\n'
-                            '`=channel freeze`/`unfreeze` - Prevent / Allow players to send messages',
-                            inline=False)
+            embed.add_field(name  = "Staff Commands",
+                            value = '`=unregister @player` - Permanently remove player profile from the system\n'
+                                    '`=channel freeze`/`unfreeze` - Prevent / Allow players to send messages',
+                            inline = False)
     except AttributeError:
         pass # if msg is from bot
     return embed
@@ -112,99 +113,115 @@ def _registerHelp(msg):
 def _lobbyHelp(msg):
     """ Returns lobby help embed
     """
-    embed = Embed(
-        colour=Color.blurple()
-    )
-    embed.add_field(name='Lobby Commands',
-                    value='`=j` - Join the lobby\n'
-                          '`=l` - Leave the lobby\n'
-                          '`=q` - See the current lobby'
-                    , inline=False)
+    embed = Embed(colour=Color.blurple())
+    embed.add_field(name  = 'Lobby Commands',
+                    value = '`=j` - Join the lobby\n'
+                            '`=l` - Leave the lobby\n'
+                            '`=q` - See the current lobby\n'
+                            '`=i` - Display the global information prompt',
+                    inline = False)
     if isAdmin(msg.author):
-        embed.add_field(name="Staff Commands",
-                        value='`=clear` - Clear the lobby\n'
-                        '`=channel freeze`/`unfreeze` - Prevent / Allow players to send messages\n'
-                        '`=remove @player` - Remove player from lobby',
-                        inline=False)
+        embed.add_field(name  = "Staff Commands",
+                        value = '`=clear` - Clear the lobby\n'
+                                '`=channel freeze`/`unfreeze` - Prevent / Allow players to send messages\n'
+                                '`=remove @player` - Remove player from lobby',
+                        inline = False)
+    return embed
+
+def _adminHelp(msg):
+    """ Returns admin help embed
+    """
+    embed = Embed(colour=Color.blurple())
+    embed.add_field(name  = 'Debug Commands',
+                    value = '`=channel (un)freeze` - Prevent users from typing in a channel\n'
+                            '`=pog version` - Display current version and lock status\n'
+                            '`=pog (un)lock` - Prevent users from interacting with the bot (but admins still can)\n',
+                    inline = False)
+    embed.add_field(name  = 'Lobby Commands',
+                    value = '`=remove @player` - Remove the player from queue\n'
+                            '`=clear` - Clear queue\n'
+                            '`=lobby get` - Get all of the user IDs to restore the lobby\n'
+                            '`=lobby restore id id ...` - Re-add all the users back to the lobby',
+                    inline = False)
+    embed.add_field(name  = 'Player Management Commands',
+                    value = '`=timeout @player duration` - Mute the player from POF for a given time\n'
+                            '`=unregister @player` - Forcibly unregisters and removes a user from the system',
+                    inline = False)
+    embed.add_field(name  = 'Match Commands',
+                    value = '`=demote @player` - Force `=resign` the player\n'
+                            '`=sub @player` - Pick someone in queue to replace the player\n'
+                            '`=map name` - Force a map\n'
+                            '`=pog ingame` - Toggle the in-game player check',
+                    inline = False)
     return embed
 
 def _defaultHelp(msg):
     """ Returns fallback help embed
     """
-    embed = Embed(
-        colour=Color.red()
-    )
-    embed.add_field(name='No available command',
-                    value='You are not supposed to use the bot on this channel',
-                    inline=False)
+    embed = Embed(colour=Color.red())
+    embed.add_field(name  = 'No available command',
+                    value = 'You are not supposed to use the bot on this channel',
+                    inline = False)
     return embed
 
 def _mapHelp(ctx):
     """ Returns map help embed
     """
-    embed = Embed(
-        colour=Color.blurple()
-    )
+    embed = Embed(colour=Color.blurple())
     cmd = ctx.command.name
     if cmd == "pick":
         cmd = "p"
-    embed.add_field(name='Map selection commands',
-                    value=f'`={cmd} a base` - Display all the maps containing *a base* in their name\n'
-                          f'`={cmd} 3` - Chooses the map number 3 from the selection\n'
-                          f'`={cmd}` - Display the current selection or show the help',
-                    inline=False)
+    embed.add_field(name  = 'Map selection commands',
+                    value = f'`={cmd} a base` - Display all the maps containing *a base* in their name\n'
+                            f'`={cmd} 3` - Chooses the map number 3 from the selection\n'
+                            f'`={cmd}` - Display the current selection or show the help',
+                    inline = False)
     return embed
 
 def _timeoutHelp(ctx):
     """ Returns timeout help embed
     """
-    embed = Embed(
-        colour=Color.blurple()
-    )
-    embed.add_field(name='Timeout command',
-                    value='`=timeout @player 10 days` - Mute @player from POG for 10 days\n'
-                          '`=timeout @player 10 hours` - Mute @player from POG for 10 hours\n'
-                          '`=timeout @player 10 minutes` - Mute @player from POG for 10 minutes\n'
-                          '`=timeout @player remove` - Unmute @player from POG\n'
-                          '`=timeout @player` - Get info on current timeout for @player',
-                    inline=False)
+    embed = Embed(colour=Color.blurple())
+    embed.add_field(name  = 'Timeout command',
+                    value = '`=timeout @player 10 days` - Mute @player from POG for 10 days\n'
+                            '`=timeout @player 10 hours` - Mute @player from POG for 10 hours\n'
+                            '`=timeout @player 10 minutes` - Mute @player from POG for 10 minutes\n'
+                            '`=timeout @player remove` - Unmute @player from POG\n'
+                            '`=timeout @player` - Get info on current timeout for @player',
+                    inline = False)
     return embed
 
 def _mutedHelp(ctx):
     """ Returns help for muted players embed
     """
-    embed = Embed(
-        colour=Color.blurple()
-    )
-    embed.add_field(name='You are currently muted!',
-                    value='`=escape` See how long you are muted for, give back permissions if no longer muted',
-                    inline=False)
+    embed = Embed(colour=Color.blurple())
+    embed.add_field(name  = 'You are currently muted!',
+                    value = '`=escape` See how long you are muted for, give back permissions if no longer muted',
+                    inline = False)
     return embed
 
 def _matchHelp(msg):
     """ Returns match help embed
     """
-    embed = Embed(
-        colour=Color.blurple(),
-    )
-    embed.add_field(name='Match commands',
-                    value='`=m` - Display the match status and team composition\n',
-                    inline=False)
-    embed.add_field(name='Team Captain commands',
+    embed = Embed(colour=Color.blurple())
+    embed.add_field(name  = 'Match commands',
+                    value = '`=m` - Display the match status and team composition\n',
+                    inline = False)
+    embed.add_field(name  = 'Team Captain commands',
                     value = '`=p @player` - Pick a player in your team\n'
                             '`=p VS`/`NC`/`TR` - Pick a faction\n'
                             '`=p base name` - Pick the map *base name*\n'
                             '`=resign` - Resign from Team Captain position\n'
-                            '`=ready` - To toggle the ready status of your team'
-                            ,
-                    inline=False)
+                            '`=ready` - To toggle the ready status of your team',
+                    inline = False)
     if isAdmin(msg.author):
-        embed.add_field(name="Staff Commands",
+        embed.add_field(name  = "Staff Commands",
                         value = '`=clear` - Clear the match\n'
-                                '`=map base name` - Select a map\n'
+                                '`=map base name` - Force select a map\n'
                                 '`=demote @player` - Remove Team Captain position from player\n'
+                                '`=sub @player` - Pick someone in queue to replace the player\n'
                                 '`=channel freeze`/`unfreeze` - Prevent / Allow players to send messages',
-                        inline=False)
+                        inline = False)
     return embed
 
 def _account(msg, account):
@@ -226,7 +243,7 @@ def _account(msg, account):
         title='Jaeger Account',
         description=f'**MUST READ: [Account usage](https://planetsideguide.com/other/jaeger/)**'
     )
-    embed.add_field(name = "Logging details",value = desc,inline=False)
+    embed.add_field(name = "Logging details",value = desc,inline = False)
     embed.set_thumbnail(
         url="https://media.discordapp.net/attachments/739231714554937455/739522071423614996/logo_png.png")
     embed.set_footer(
@@ -244,6 +261,8 @@ def _autoHelp(msg):
         return _matchHelp(msg)
     if msg.channel.id == cfg.channels['muted']:
         return _mutedHelp(msg)
+    if msg.channel.id == cfg.channels['staff']:
+        return _adminHelp(msg)
     return _defaultHelp(msg)
 
 def _lobbyList(msg, namesInLobby):
@@ -253,7 +272,7 @@ def _lobbyList(msg, namesInLobby):
     listOfNames="\n".join(namesInLobby)
     if listOfNames == "":
         listOfNames = "Queue is empty"
-    embed.add_field(name=f'Lobby: {len(namesInLobby)} / {cfg.general["lobby_size"]}',value=listOfNames,inline=False)
+    embed.add_field(name=f'Lobby: {len(namesInLobby)} / {cfg.general["lobby_size"]}',value=listOfNames,inline = False)
     return embed
 
 
@@ -262,7 +281,7 @@ def _selectedMaps(msg, sel):
     """
     embed = Embed(colour=Color.blue())
     maps = sel.stringList
-    embed.add_field(name=f"{len(maps)} maps found",value="\n".join(maps),inline=False)
+    embed.add_field(name=f"{len(maps)} maps found",value="\n".join(maps),inline = False)
     return embed
 
 def _offlineList(msg, pList):
@@ -271,11 +290,38 @@ def _offlineList(msg, pList):
         title='Offline Players',
         description=f'If your character info is incorrect, re-register using `=r` in <#{cfg.channels["register"]}>!'
     )
-    embed.add_field(name="The following players are not online ingame:",
+    embed.add_field(name  = "The following players are not online ingame:",
                     value = "\n".join(f"{p.igName} ({p.mention})" for p in pList),
-                    inline=False)
+                    inline = False)
 
     return embed
+
+
+def _globalInfo(msg, lobby, matchList):
+    embed = Embed(
+        colour=Color.greyple(),
+        title='Global Info',
+        description=f'POG bot version `{cfg.VERSION}`'
+    )
+    lbEmbed = _lobbyList(msg, namesInLobby=lobby).fields[0]
+    embed.add_field(name = lbEmbed.name, value = lbEmbed.value, inline = lbEmbed.inline)
+    for m in matchList:
+        desc = ""
+        if m.status is not MatchStatus.IS_FREE:
+            if m.roundNo != 0:
+                desc += f"*Match {m.number} - Round {m.roundNo}*"
+            else:
+                desc += f"*Match {m.number}*"
+            desc += "\n"
+        desc += f"Status: {m.statusString}"
+        if m.status in (MatchStatus.IS_WAITING, MatchStatus.IS_PLAYING, MatchStatus.IS_RESULT):
+            desc += f"\nBase: **{m.map.name}**\n"
+            desc += " / ".join(f"{tm.name}: **{cfg.factions[tm.faction]}**" for tm in m.teams)
+        if m.status is MatchStatus.IS_PLAYING:
+            desc += f"\nTime Remaining: **{m.formatedTimeToRoundEnd}**"
+        embed.add_field(name  = m.channel.name, value = desc, inline = False)
+    return embed
+
 
 def _teamUpdate(arg, match):
     """ Returns the current teams
@@ -290,9 +336,12 @@ def _teamUpdate(arg, match):
         title = f"Match {match.number} - Round {match.roundNo}"
     else:
         title = f"Match {match.number}"
-    embed = Embed(colour=Color.blue(), title=title, description = match.statusString)
+    desc = match.statusString
+    if match.status is MatchStatus.IS_PLAYING:
+        desc += f"\nTime Remaining: **{match.formatedTimeToRoundEnd}**"
+    embed = Embed(colour=Color.blue(), title=title, description = desc)
     if match.map is not None:
-        embed.add_field(name = "Map",value = match.map.name,inline=False)
+        embed.add_field(name = "Map",value = match.map.name,inline = False)
     for tm in match.teams:
         value = ""
         name = ""
@@ -312,9 +361,9 @@ def _teamUpdate(arg, match):
             name = f"{tm.name}"
         embed.add_field(name = name,
                         value = value,
-                        inline=False)
+                        inline = False)
     if match.status is MatchStatus.IS_PICKING:
-        embed.add_field(name=f'Remaining',value="\n".join(match.playerPings) ,inline=False)
+        embed.add_field(name=f'Remaining',value="\n".join(match.playerPings) ,inline = False)
     return embed
 
 def _jaegerCalendar(arg):
@@ -323,7 +372,7 @@ def _jaegerCalendar(arg):
     date = dt.now(tz.utc)
     embed.add_field(name = "Current UTC time",
                         value = date.strftime("%Y-%m-%d %H:%M UTC"),
-                        inline=False)
+                        inline = False)
     return embed
 
 class _Message():
@@ -445,6 +494,7 @@ class _StringEnum(Enum):
     CHANNEL_INIT = _Message("`Bot init`: Correctly hooked in channel <#{}>")
     INVALID_STR = _Message("You entered an invalid caracter! `{}`")
     API_ERROR = _Message("Could not reach Planetside2 API, try again later!")
+    GLOBAL_INFO = _Message("Here is what's going on in POG at the moment:", embed=_globalInfo)
 
     BOT_UNLOCKED = _Message("Unlocked!")
     BOT_LOCKED = _Message("Locked!")
@@ -521,4 +571,11 @@ class _StringEnum(Enum):
     MUTE_SHOW = _Message("You are muted from POG until {}!")
     MUTE_FREED = _Message("You are no longer muted from POG!")
 
-    SC_ILLEGAL_WE = _Message("Alert: {} used {} during match {}!")
+    SC_ILLEGAL_WE = _Message("{} used {} during match {}! This weapon is banned! Ignoring {} kill(s)...")
+
+    SUB_NO = _Message("This player can't be subbed!")
+    SUB_NO_CAPTAIN = _Message("Can't sub a Team Captain!")
+    SUB_NO_PLAYER = _Message("No player is available to substitute!")
+    SUB_OKAY_TEAM = _Message("{} replaced {} in {}", ping=False, embed=_teamUpdate)
+    SUB_OKAY = _Message("{} replaced {}!", ping=False, embed=_teamUpdate)
+    SUB_LOBBY = _Message("{} you have been designed as a substitute, join <#{}>!", embed=_lobbyList)